requirements.txt
deepface==0.0.79
facenet_pytorch==2.5.3
Flask==3.0.0
Flask_Cors==4.0.0
Pillow==10.1.0
PyJWT==2.8.0
PyJWT==2.8.0
torch==2.1.0
torchvision==0.16.0
Werkzeug==3.0.0
flask_sqlalchemy==3.1.1
psycopg2==2.9.9